port

+++ code/modules/mob/living/simple_animal/friendly/farm_animals.dm
@@ -147,26 +147,23 @@
 	return ..()
 
 /mob/living/simple_animal/cow/attackby(obj/item/O, mob/user, params)
-	if(stat == CONSCIOUS)
-		if(istype(O, /obj/item/reagent_containers/glass))
-			udder.milkAnimal(O, user)
-			return 1
-		else if(istype(O, food_type))
-			if(is_calf)
-				visible_message("<span class='alertalien'>[src] adorably chews the [O].</span>")
-				qdel(O)
-				return 1
-			else if(!has_calf)
-				has_calf = 1
-				visible_message("<span class='alertalien'>[src] hungrily consumes the [O].</span>")
-				qdel(O)
-				return 1
-			else
-				visible_message("<span class='alertalien'>[src] absently munches the [O].</span>")
-				qdel(O)
-				return 1
+	if(stat == CONSCIOUS && istype(O, /obj/item/reagent_containers/glass)) // Should probably be bound into a proc at this point.
+		udder.milkAnimal(O, user)
+		return 1
+	if(stat == CONSCIOUS && istype(O, food_type))
+		if(is_calf)
+			visible_message("<span class='alertalien'>[src] adorably chews the [O].</span>")
+			qdel(O)
+		if(!has_calf && !is_calf)
+			has_calf = 1
+			visible_message("<span class='alertalien'>[src] hungrily consumes the [O].</span>")
+			qdel(O)
+		else
+			visible_message("<span class='alertalien'>[src] absently munches the [O].</span>")
+			qdel(O)
 	else
 		return ..()
+
 
 /mob/living/simple_animal/cow/Life()
 	. = ..()
@@ -182,7 +179,7 @@
 			if((prob(3)))
 				is_calf = 0
 				udder = new()
-				if (name == "brahmin calf") 
+				if (name == "brahmin calf")
 					name = "brahmin"
 				else
 					name = "cow"
@@ -267,7 +264,7 @@
 
 /mob/living/simple_animal/chicken
 	name = "\improper chicken"
-	desc = "Hopefully the eggs are good this season."
+	desc = "One of the two kinds the legion prefers."
 	gender = FEMALE
 	mob_biotypes = list(MOB_ORGANIC, MOB_BEAST)
 	icon_state = "chicken_brown"
@@ -382,7 +379,7 @@
 
 /mob/living/simple_animal/cow/brahmin
 	name = "brahmin"
-	desc = "Brahmin or brahma are mutated cattle with two heads and giant udders.<br>Known for their milk, just don't tip them over."
+	desc = "Brahmin or brahma are mutated cattle with two heads and looking udderly ridiculous.<br>Known for their milk, just don't tip them over."
 	icon = 'icons/mob/wastemobs.dmi'
 	icon_state = "brahmin"
 	icon_living = "brahmin"
@@ -397,7 +394,7 @@
 
 /mob/living/simple_animal/hostile/retaliate/goat/bighorn
 	name = "big horner"
-	desc = "Mutated bighorn sheep that are often found in mountains. You should start moving away from them If they stare at you"
+	desc = "Mutated bighorn sheep that are often found in mountains, and are known for being foul-tempered even at the best of times."
 	icon = 'icons/mob/wastemobs.dmi'
 	icon_state = "bighorner"
 	icon_living = "bighorner"
@@ -432,26 +429,22 @@
 	var/young_type = /mob/living/simple_animal/hostile/retaliate/goat/bighorn/calf
 	stop_automated_movement_when_pulled = 1
 	blood_volume = BLOOD_VOLUME_NORMAL
-	
+
 /mob/living/simple_animal/hostile/retaliate/goat/bighorn/attackby(obj/item/O, mob/user, params)
-	if(stat == CONSCIOUS)
-		if(istype(O, /obj/item/reagent_containers/glass))
-			udder.milkAnimal(O, user)
-			return 1
-		else if(istype(O, food_type))
-			if(is_calf)
-				visible_message("<span class='alertalien'>[src] adorably chews the [O].</span>")
-				qdel(O)
-				return 1
-			else if(!has_calf)
-				has_calf = 1
-				visible_message("<span class='alertalien'>[src] hungrily consumes the [O].</span>")
-				qdel(O)
-				return 1
-			else
-				visible_message("<span class='alertalien'>[src] absently munches the [O].</span>")
-				qdel(O)
-				return 1
+	if(stat == CONSCIOUS && istype(O, /obj/item/reagent_containers/glass)) // Should probably be bound into a proc at this point.
+		udder.milkAnimal(O, user)
+		return 1
+	if(stat == CONSCIOUS && istype(O, food_type))
+		if(is_calf)
+			visible_message("<span class='alertalien'>[src] adorably chews the [O].</span>")
+			qdel(O)
+		if(!has_calf && !is_calf)
+			has_calf = 1
+			visible_message("<span class='alertalien'>[src] hungrily consumes the [O].</span>")
+			qdel(O)
+		else
+			visible_message("<span class='alertalien'>[src] absently munches the [O].</span>")
+			qdel(O)
 	else
 		return ..()
 
@@ -469,24 +462,34 @@
 			if((prob(3)))
 				is_calf = 0
 				udder = new()
-				if (name == "bighorn lamb") 
+				if (name == "bighorn lamb")
 					name = "bighorn"
 				else
 					name = "bighorn"
 				visible_message("<span class='alertalien'>[src] has fully grown.</span>")
 		else
 			udder.generateMilk()
-	
+
 /mob/living/simple_animal/hostile/retaliate/goat/bighorn/calf
 	name = "bighoner lamb"
 	resize = 0.55
 
+/mob/living/simple_animal/hostile/retaliate/goat/bighorn/calf/Initialize() //calfs should not be a separate critter, they should just be a normal whatever with these vars
+	. = ..()
+	resize = 0.55
+
 /mob/living/simple_animal/cow/calf
 	name = "cow calf"
-	resize = 0.55
 	is_calf = 1
+
+/mob/living/simple_animal/cow/calf/Initialize()
+	. = ..()
+	resize = 0.55
 
 /mob/living/simple_animal/cow/brahmin/calf
 	name = "brahmin calf"
-	resize = 0.55
 	is_calf = 1
+
+/mob/living/simple_animal/cow/brahmin/calf/Initialize()
+	. = ..()
+	resize = 0.55
